Squashed commit of the following:
commit b1ff0063b3e332ab953934275b68fd95ba8cd8bf
Merge: 4ec6792 f09b0c7
Author: Hemant Jam <60656624+hemantjamdev@users.noreply.github.com>
Date:   Sat May 13 17:10:48 2023 +0530

    Merge branch 'main' into google_backup

commit 4ec6792cb37ecc068375a232fed2885238d83777
Author: Hemant Jam <60656624+hemantjamdev@users.noreply.github.com>
Date:   Sat May 13 17:07:02 2023 +0530

    google backup

commit 2d6b8e264f92f5418f71e091c062bbfcd9deb001
Author: Hemant Jam <60656624+hemantjamdev@users.noreply.github.com>
Date:   Fri May 12 18:26:39 2023 +0530

    back-up

commit 8578a56193b2c4097a80c97b1c87cd1685a455aa
Author: Hemant Jam <60656624+hemantjamdev@users.noreply.github.com>
Date:   Fri May 12 18:25:25 2023 +0530

    google signin

commit c8dc995c657505fc31ab33f21dbcdd2f63253152
Author: Hemant Jam <60656624+hemantjamdev@users.noreply.github.com>
Date:   Fri May 12 18:25:03 2023 +0530

    google signin

+++ lib/widgets/delete_all.dart
@@ -1,18 +1,24 @@
 import 'package:flutter/material.dart';
-import 'package:note/constants/app_icons.dart';
+import 'package:note/constants/strings.dart';
 import 'package:note/local_storage/local_storage.dart';
 import 'package:note/widgets/confirmation_dialog.dart';
 
-Widget buildDeleteAll({required BuildContext context, required String text}) {
+Widget buildDeleteAll({required BuildContext context}) {
   DBHelper dbHelper = DBHelper();
   return IconButton(
-    icon:Icon(Icons.delete,),
-    onPressed: () {
-      confirmation(context: context, text: text).then((value) {
-        if (value) {
+    icon: const Icon(Icons.delete),
+    onPressed: () async {
+      if (context.mounted) {
+        bool? value = await confirmation(
+            content: Strings.areYourSure,
+            context: context,
+            title: Strings.delete,
+            cancel: Strings.cancel,
+            confirm: Strings.confirm);
+        if (value != null && value) {
           dbHelper.deleteAll();
         }
-      });
+      }
     },
   );
 }
